v2.4.1: fix int32 overflow in newBoards counter
WorkerLevelStats::newBoards was std::atomic<int> (signed 32-bit, max
~2.1 B); at level 14 of the 6x6 solve (~2.87 B new boards) it
overflowed to a large negative value.  The error cascaded into the Pass
column (totalChildren - newBoardsOut), printing an impossibly large
positive number.  Widened to std::atomic<long long> to match
totalChildren and terminalBoards.

Also removed a spurious (int) cast on passBatch.size() that masked the
same class of overflow when accumulating into long long boardsIn.

+++ OthelloSolverCommandLine/SolverWorker.h
@@ -10,7 +10,7 @@
 struct WorkerLevelStats
 {
     std::atomic<int>        activeCount{0};    // in-flight batch count
-    std::atomic<int>        newBoards{0};      // new unique child boards inserted
+    std::atomic<long long>  newBoards{0};      // new unique child boards inserted
     std::atomic<long long>  totalChildren{0};  // total children generated (new + dups)
     std::atomic<long long>  terminalBoards{0}; // boards with no legal moves (game over)
     std::atomic<int>        maxMovesInLevel{0};// max legal moves seen for any board at this level
